backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==3.14.0
django-filter==24.3
celery>=5.3.0
redis>=4.0.0
asyncio>=3.4.3
aioredis>=2.0.1
dnspython>=2.7.0
pyOpenSSL>=24.3.0
djangorestframework-simplejwt==5.3.0
django-elasticsearch-dsl==8.0
elasticsearch==8.17.0
elasticsearch-dsl==8.17.0
